Indexing code cleanup.
